Add files via upload
In this project, I worked with data from the CIA World Factbook, a compendium of statistics about all of the countries on Earth. The Factbook contains demographic information like:

population - The population as of 2015.
population_growth - The annual population growth rate, as a percentage.
area - The total land and water area.

I'll explore the Python SQLite workflow to explore, analyze, and visualize data from this database.